+++ app/rest/server.go
@@ -2,60 +2,52 @@ package rest
 
 import (
 	"time"
-	"net/http"
 	"log"
 	"fmt"
 	"strconv"
-	"encoding/json"
 
+	"github.com/gin-gonic/gin"
 	"github.com/popstas/pixel-server/app/pixel"
 	"github.com/popstas/pixel-server/app/kapacitor"
 )
 
 type Server struct {
-	Pixel         pixel.Pixel
-	HostPort      string
+	HostPort string
+	Pixel    pixel.Pixel
 }
 
-func (ps Server) Run() {
-	ps.Pixel.SetStatus(pixel.PixelData{ 100, fmt.Sprintf("server started\\%s", ps.HostPort), 1, 20 })
+func (s Server) Run() {
+	gin.SetMode(gin.ReleaseMode)
+	router := gin.Default()
+
+	router.POST("/status", s.statusHandler)
+	router.POST("/kapacitor", s.kapacitorHandler)
+
+	s.Pixel.SetStatus(pixel.PixelData{ 100, fmt.Sprintf("server started\\%s", s.HostPort), 1, 20 })
 	time.Sleep(2000 * time.Millisecond)
-	ps.Pixel.SetStatus(pixel.PixelData{ -1, "", 0, 100 })
+	s.Pixel.SetStatus(pixel.PixelData{ -1, "", 0, 100 })
 
-	http.HandleFunc("/status", ps.statusHandler)
-	http.HandleFunc("/kapacitor", ps.kapacitorHandler)
-	http.ListenAndServe(ps.HostPort, nil)
+	log.Fatal(router.Run(s.HostPort))
 }
 
-func (ps *Server) statusHandler(w http.ResponseWriter, r *http.Request) {
-	if(r.Method != "POST") {
-		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
-		return
-	}
-	r.ParseForm()
-
+func (s *Server) statusHandler(c *gin.Context) {
 	var pd pixel.PixelData
 	var err error
 
-	pd.Value, _ = strconv.Atoi(r.FormValue("value"))
-	pd.Message = r.FormValue("message")
-	pd.Blink, _ = strconv.Atoi(r.FormValue("blink"))
-	pd.Brightness, err = strconv.Atoi(r.FormValue("brightness"))
+	pd.Value, _ = strconv.Atoi(c.PostForm("value"))
+	pd.Message = c.PostForm("message")
+	pd.Blink, _ = strconv.Atoi(c.PostForm("blink"))
+	pd.Brightness, err = strconv.Atoi(c.PostForm("brightness"))
 	if err != nil{
 		pd.Brightness = 100
 	}
 
-	go ps.Pixel.SetStatus(pd)
+	go s.Pixel.SetStatus(pd)
 }
 
-func (ps Server) kapacitorHandler(w http.ResponseWriter, r *http.Request) {
-	if(r.Method != "POST") {
-		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
-		return
-	}
+func (s Server) kapacitorHandler(c *gin.Context) {
 	ad := kapacitor.KapacitorAlertData{}
-	decoder := json.NewDecoder(r.Body)
-	err := decoder.Decode(&ad)
+	err := c.BindJSON(ad)
 	if err != nil {
 		log.Fatalf("Could not decode kapacitor AlertData, %s", err)
 	}
@@ -79,6 +71,5 @@ func (ps Server) kapacitorHandler(w http.ResponseWriter, r *http.Request) {
 	data := ad.Data.Series[0]
 	pd.Message = fmt.Sprintf("%s\\%s: %v", data.Tags.Host, data.Name, data.Values[0][1]) // data.Values[1]
 
-	go ps.Pixel.SetStatus(pd)
+	go s.Pixel.SetStatus(pd)
 }
-