+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ## NEXT VERSION
 
 FEATURES:
-  * receive status to /status?value=50&message=msg\second line&blink=1&brightness=100
-  * receive status from kapacitor post to /kapacitor
 
 IMPROVEMENTS:
+  * code decoupled to packages
+  * server rewrited from net/http to gin-gonic/gin
+  * add release binaries for windows and linux
 
 BUG FIXES:
+
+## v0.1 (30.08.2016)
+
+FEATURES:
+  * receive status to `/status?value=50&message=msg\second line&blink=1&brightness=100`
+  * receive status from kapacitor post to `/kapacitor`